causalmlr: Causal Machine Learning with mlr3

Estimation and evaluation of average and conditional average treatment effects (ATE/CATE) from observational data using machine learning, built on the mlr3 ecosystem.

ATE estimators

ate_naive() - difference in means (baseline)

ate_ipw() - inverse propensity weighting

ate_dr() - doubly robust / AIPW

ate_dml() - double machine learning (partially linear model)

CATE meta-learners

s_learner(), t_learner(), x_learner(), dr_learner()

Evaluation

eps_ate() - absolute ATE error (needs ground truth)

pehe() - precision in estimating heterogeneous effects (needs ground truth)

r_loss() with rloss_nuisance() - observable score for model selection and hyperparameter tuning on real data

All estimators accept arbitrary mlr3 learners as nuisance models, so anything from mlr3::lrn("regr.lm") to a tuned AutoTuner can be plugged in.

See also

Author

Maintainer: Damian Machlanski d.machlanski@ed.ac.uk

Authors:

Damian Machlanski d.machlanski@ed.ac.uk