IHDP semi-synthetic benchmark (training set)

The Infant Health and Development Program (IHDP) benchmark introduced by Hill (2011): real covariates describing children and their mothers from a randomised experiment, combined with a simulated outcome, which provides ground-truth treatment effects. The treatment is intensive, high-quality childcare and home visits; the (simulated) outcome mimics future cognitive test scores.

ihdp_train

Format

A data frame with 672 rows and 28 columns:

x0-x5

Continuous covariates (standardised).

x6-x24

Binary covariates.

t

Binary treatment indicator (0/1).

y

Continuous (simulated) outcome.

tau

True conditional average treatment effect.

References

Hill, J. L. (2011). Bayesian nonparametric modeling for causal inference. Journal of Computational and Graphical Statistics, 20(1), 217-240.

Examples

data(ihdp_train)
head(ihdp_train)
#>           x0        x1          x2         x3          x4         x5 x6 x7 x8
#> 1  1.3973950 0.9963463 -1.10562395 -0.8796060  0.30856884 -1.0234024  1  0  0
#> 2  0.2690327 0.1968181  0.38382797  0.1617025 -0.62918919  1.4608315  1  0  1
#> 3  1.0515368 1.7958744 -1.10562395  0.1617025 -0.62918919  0.9639847  1  0  1
#> 4  0.6624464 0.1968181 -0.73326097 -0.8796060  0.37108604 -0.6921712  1  0  0
#> 5  0.8569916 1.7958744  0.01146499 -0.8796060  0.55863765  0.3015224  0  1  1
#> 6 -0.3535121 0.1968181  0.75619095 -0.8796060 -0.00401717 -1.3546336  0  0  0
#>   x9 x10 x11 x12 x13 x14 x15 x16 x17 x18 x19 x20 x21 x22 x23 x24 t         y
#> 1  0   1   0   1   1   0   0   1   1   0   0   0   0   0   0   1 1 4.7712323
#> 2  0   0   1   0   0   0   1   1   1   1   0   0   0   0   0   0 0 2.9562729
#> 3  1   0   0   0   0   0   1   1   1   0   0   0   0   0   0   1 0 4.1641645
#> 4  0   0   1   0   1   0   1   0   1   0   1   0   0   0   0   0 1 6.1723073
#> 5  0   0   0   0   1   0   1   0   1   1   0   0   0   0   0   0 1 7.8344689
#> 6  1   0   0   0   1   0   1   0   1   0   0   0   0   0   0   0 0 0.7537379
#>        tau
#> 1 4.657928
#> 2 3.428604
#> 3 3.658195
#> 4 4.585505
#> 5 4.265591
#> 6 4.662679