Jobs benchmark (training set)

The Jobs benchmark (LaLonde, 1986; composition of Shalit et al., 2017) combines a randomised experiment (the National Supported Work programme) with observational controls. The treatment is participation in a job training programme and the outcome is employment status. The e column flags units belonging to the randomised subsample, which enables evaluation on real data via the experimental subset.

jobs_train

Format

A data frame with 2,891 rows and 20 columns:

x0-x16

Covariates (standardised continuous and binary).

t

Binary treatment indicator: job training (0/1).

y

Binary outcome: employment (0/1).

e

Indicator of membership in the randomised experimental subsample (0/1).

References

LaLonde, R. J. (1986). Evaluating the econometric evaluations of training programs with experimental data. The American Economic Review, 76(4), 604-620.

Shalit, U., Johansson, F. D., & Sontag, D. (2017). Estimating individual treatment effect: generalization bounds and algorithms. Proceedings of the 34th International Conference on Machine Learning.

Examples

data(jobs_train)
head(jobs_train)
#>           x0         x1 x2 x3 x4 x5          x6         x7          x8
#> 1 -0.6142821  1.4647266  0  0  1  0  2.39325000  2.7461965 -0.65331121
#> 2 -0.8024627  0.1018350  0  0  1  0  0.10988519  0.4982713 -0.78528431
#> 3 -0.8965531 -0.2388879  1  0  1  1 -0.08521244 -0.1480969 -0.84731167
#> 4 -0.8965531 -0.2388879  0  0  0  1  0.40558116  0.3255935 -0.84731167
#> 5  0.1384405 -1.6017795  1  0  1  1 -0.72253135 -0.2127338 -0.01984034
#> 6  0.1384405  0.4425579  0  0  1  0  2.16202319  1.9130996 -0.01984034
#>           x9         x10         x11         x12 x13 x14         x15 x16 t y e
#> 1 -0.6569127  1.62753104  2.46233663  2.93724407   0   0  2.84390896   0 0 1 0
#> 2 -0.7434072 -0.02250229 -0.17719290  0.08253676   0   0  0.03842226   0 0 1 0
#> 3 -0.7816058 -0.36134842 -0.28622144 -0.30361510   0   0 -0.19130379   0 0 1 0
#> 4 -0.7816058 -0.36134842  0.02301955 -0.03962994   0   0  0.17310822   0 0 1 1
#> 5 -0.1560193 -1.42208413 -0.51456318 -0.33155179   0   0 -0.77922690   0 0 1 0
#> 6 -0.1560193  0.34580872  2.08071790  1.60539169   0   0  1.92402928   0 0 1 0