Sodium intake and blood pressure (synthetic)

A simplified simulation of the effect of sodium intake on blood pressure, proposed by Luque-Fernandez et al. (2019). Age confounds both the treatment (high sodium intake) and the outcome (blood pressure). The true ATE used in the data generating process is 1.05.

sodium

Format

A data frame with 10,000 rows and 3 columns:

age

Age in years (confounder).

sodium

Binary treatment: high sodium intake (1) or not (0).

bp

Systolic blood pressure (outcome).

References

Luque-Fernandez, M. A., Schomaker, M., Redondo-Sanchez, D., Jose Sanchez Perez, M., Vaidya, A., & Schnitzer, M. E. (2019). Educational Note: Paradoxical collider effect in the analysis of non-communicable disease epidemiological data. International Journal of Epidemiology, 48(2), 640-653.

Examples

data(sodium)
ate_naive(sodium, outcome = "bp", treatment = "sodium")
#> ATE estimate - Naive (difference in means) 
#>   Estimate:  5.18
#>   Std. error: 0.1948 
#>   95% CI:    [4.798, 5.561]
#>   N:         10000