Synthetic CATE benchmark (training set)

Fully synthetic data generated with the data generating process of Künzel et al. (2019), with five Gaussian covariates, a binary treatment and a continuous outcome. The true CATE is a step function of one of the covariates. See synth_test for the accompanying test set with ground truth effects.

synth_train

Format

A data frame with 1,000 rows and 7 columns:

x0, x1, x2, x3, x4

Continuous covariates.

t

Binary treatment indicator (0/1).

y

Continuous outcome.

References

Künzel, S. R., Sekhon, J. S., Bickel, P. J., & Yu, B. (2019). Metalearners for estimating heterogeneous treatment effects using machine learning. Proceedings of the National Academy of Sciences, 116(10), 4156-4165.

Examples

data(synth_train)
head(synth_train)
#>           x0          x1          x2         x3         x4 t          y
#> 1 -2.1966546  1.35390014 -1.39801445 -1.7138880 -0.6693249 1 12.1009267
#> 2  0.6193071 -0.03483849 -0.03802601  0.1585317  1.0855784 1  0.4292407
#> 3  0.1146665  0.57386661  0.53158533 -1.3231972 -1.3461593 0  1.0622299
#> 4 -0.7829745  1.20219763 -1.04630459 -1.0525001  0.2200478 0  2.4699487
#> 5  0.8731837  1.32112750  0.22711515 -0.5524267  1.0754895 1  8.4183230
#> 6  2.3052394  0.32871172  0.20735839  1.1505775  1.4527570 1  8.5148841